Public export of Nexus code:   - d682adfbf5e79b7677ee92fbeb0012f99aa6d795 by Michael Kearns <mkearns@sonatype.com>
GitOrigin-RevId: d682adfbf5e79b7677ee92fbeb0012f99aa6d795

+++ components/nexus-repository-content/src/main/java/org/sonatype/nexus/repository/content/browse/store/BrowseNodeManager.java
@@ -93,7 +93,7 @@ public class BrowseNodeManager
    * Creates browse nodes for the path (runs in a single transaction).
    */
   protected void doCreateBrowseNodes(final List<BrowsePath> paths, final Consumer<BrowseNodeData> finalStep) {
-    Integer parentId = null;
+    Long parentId = null;
     for (int i = 0; i < paths.size(); i++) {
       BrowseNodeData node = new BrowseNodeData();
       node.setRepositoryId(repositoryId);
